EP10-Real-world/flask-app/requirements.txt
flask
psycopg2-binary

